+++ Back/Banco-de-Talentos/src/main/java/com/fractal/bancodetalentos/model/response/RolResp.java
@@ -9,7 +9,7 @@ import java.math.BigDecimal;
 @AllArgsConstructor
 @NoArgsConstructor
 public class RolResp {
-    private BigDecimal id;
+    private Integer id;
     private String name;
     private String code;
 }
